CommonDatat 수정

+++ TextRPG_18/CommonData.cs
@@ -37,13 +37,8 @@ enum ItemType
 
 struct playermax
 {
-    public static int atk;
+    public static float atk;
     public static int dfs;
     public static int CRP;
     public static int CRD;
 }
-
-//struct playerConst
-//{
-//    public const int maxHp = 100;
-//}
